% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{calc1}
\alias{calc1}
\title{wikipedia formula https://es.wikipedia.org/wiki/Punto_de_roc%C3%ADo
RH relative humidity in % (values from 0 to 100)
temp temperature in 째C}
\usage{
calc1(RH, temp)
}
\arguments{
\item{RH}{[in percentage] an integer or double value between 0 and 100.}

\item{temp}{[째C] an integer or double value between -20 and 60 째C}
}
\value{
dew point value (double)
}
\description{
wikipedia formula https://es.wikipedia.org/wiki/Punto_de_roc%C3%ADo
RH relative humidity in % (values from 0 to 100)
temp temperature in 째C
}
