% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{calc2}
\alias{calc2}
\title{Mark G. Lawrence approach
https://journals.ametsoc.org/doi/pdf/10.1175/BAMS-86-2-225
The Relationship between Relative Humidity and the Dewpoint Temperature in Moist Air:
A Simple Conversion and Applications
https://doi.org/10.1175/BAMS-86-2-225}
\usage{
calc2(RH, temp)
}
\arguments{
\item{RH}{[in percentage] an integer or double value between 0 and 100.}

\item{temp}{[°C] an integer or double value between -20 and 60 °C}
}
\value{
dew point value (double)
}
\description{
Mark G. Lawrence approach
https://journals.ametsoc.org/doi/pdf/10.1175/BAMS-86-2-225
The Relationship between Relative Humidity and the Dewpoint Temperature in Moist Air:
A Simple Conversion and Applications
https://doi.org/10.1175/BAMS-86-2-225
}
