% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{calc3}
\alias{calc3}
\title{Alduchov and Eskridge (1996)
paper: Alduchov, O. A., and R. E. Eskridge, 1996:
Improved Magnus' form approximation of saturation vapor pressure. J. Appl. Meteor., 35, 601–609.}
\usage{
calc3(RH, temp)
}
\arguments{
\item{RH}{[in percentage] an integer or double value between 0 and 100.}

\item{temp}{[°C] an integer or double value between -20 and 60 °C}
}
\value{
dew point value (double)
}
\description{
Alduchov and Eskridge (1996)
paper: Alduchov, O. A., and R. E. Eskridge, 1996:
Improved Magnus' form approximation of saturation vapor pressure. J. Appl. Meteor., 35, 601–609.
}
