% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{calcDewPoint}
\alias{calcDewPoint}
\title{DEW POINT estimation given relative humidity and temperature}
\usage{
calcDewPoint(RH, temp, mode = "A")
}
\arguments{
\item{RH}{[in percentage] an integer or double value between 0 and 100.}

\item{temp}{[°C] an integer or double value between -20 and 60 °C.}

\item{mode}{string values "A", "B" or "C". Default "A".}
}
\value{
dew point value (double)
}
\description{
Mode "A" : calls calcDewPoint.A function
Mode "B" : calls calcDewPoint.B function
Mode "C":  calls calcDewPoint.C function
}
