% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{convert.temperature}
\alias{convert.temperature}
\title{Temperature conversion}
\usage{
convert.temperature(from = "F", to = "C", values)
}
\arguments{
\item{from}{possible values, "F" for Fahrenheit, "C" degress Celsius and "K" Kelvin.}

\item{to}{possible values, "F" for Fahrenheit, "C" degress Celsius and "K" Kelvin.}

\item{values}{can be an vector, array, or a numeric single value.}
}
\value{
value (double)
}
\description{
Temperature conversion from/to Fahrenheit (°F), degress Celsius (°C) and Kelvin (K)
}
\examples{
library(frost)
convert.temperature(from="K", to="C",350)
cels <- convert.temperature(from="F",to="C",c(120,80,134,110))
k <- convert.temperature(from="C", to="K",cels)

}
