% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ref.r
\name{is.ref}
\alias{is.ref}
\title{Whether or not an object is a frost reference.}
\usage{
is.ref(x)
}
\arguments{
\item{x}{ANY. Any R object.}
}
\value{
whether or not it is a frost reference of S3 class
   \code{"frostRef"}.
}
\description{
Whether or not an object is a frost reference.
}

