% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predTrend}
\alias{predTrend}
\title{Predict the trend of the temperature during a frost nigth.
This equation has been taken
from UC Davis formula http://biomet.ucdavis.edu/frostprotection/fp002.htm}
\usage{
predTrend(Tmin, t2, i, n)
}
\arguments{
\item{Tmin}{predicted minimum temperature}

\item{t2}{temperature 2 hours after sunset}

\item{i}{value which must be i < n, i hours value after sunset}

\item{n}{how many hours for sunrise}
}
\value{
An R basic scatter/line plot
}
\description{
Predict the trend of the temperature during a frost nigth.
This equation has been taken
from UC Davis formula http://biomet.ucdavis.edu/frostprotection/fp002.htm
}
\examples{
print("ACA VA UN EJEMPLO")
}
