% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{square}
\alias{square}
\title{Square a number}
\usage{
square(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
The square of the input
}
\description{
Takes any number and squares it
}
\examples{
print(square(4.5))
}
