% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{predFAO}
\alias{predFAO}
\title{Predict the minimum temperature using the recommended FAO equation}
\usage{
predFAO(model, t, dw = NULL)
}
\arguments{
\item{model}{: a FAOFrostModel object}

\item{t}{[째C]: an array of ambient temperature, two hours after sunset.}

\item{dw}{[째C]: an array of dew points, two hours after sunset, by default is NULL.}
}
\value{
tmin [째C]: minimum temperature
}
\description{
Predict the minimum temperature using the recommended FAO equation, which
can be applied to nights with radiative frost (wind less than 2 m/s^2, no clouds, no fog, no rain).
}
\details{
The method was extracted from the documentation and previous implementation in FFST.xls file,
which is name in the book
"Frost Protection: fundamentals, practice, and economics. Volume 1.
Authors: Richard L Snyder, J. Paulo de Melo-Abreu.
Food and Agriculture Organization of the United Nations. 2005"

This function returns Tmim = a * temp + b * dew + c
if dew argument is not null.
Otherwise, return Tmin = a * temp + c, if dew is NULL.

For more details please check:
<http://www.fao.org/docrep/008/y7223e/y7223e0b.htm#bm11.8>
<http://www.fao.org/docrep/008/y7223e/y7223e0b.htm>
FFST spreasheet <http://biomet.ucdavis.edu/frostprotection/FTrend/FFST_FTrend.htm>
}
\examples{
t0 <- c(3.2,0.8,0.2,2.6,4.4,5.2,2.7,1.2,4.5,5.6) # temperature 2 hours after sunset
td <- c(-4.2,-8.8,-6.5,-6.2,-6.1,2.6,-0.7,-1.7,-1.2,0.1) # dew point 2 hours after sunset
tn <- c(-3.1,-5,-6.3,-5.4,-4,-2.5,-4.8,-5,-4.4,-3.3)
out <- buildFAO(dw = td,temp=t0,tmin=tn)
current_temp <- 10
current_dw <- 2
ptmin <- predFAO(out,current_temp,current_dw)
cat("The predicte minimum temperature is ",ptmin," 째C")

}
