% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ref.r
\docType{class}
\name{ref}
\alias{ref}
\title{name ref}
\usage{
ref(x)
}
\arguments{
\item{x}{object. This will be converted to a reference. That is,
if character, it will be wrapped in a \code{frostRef} class,
and if not, it will be deparsed and substituted and wrapped
in a \code{frostRef} class.}
}
\description{
name ref
}

